varnish cookie check moved to cookie module
