+++ test/spec/spid-cookie_test.js
@@ -1,8 +1,8 @@
 describe('SPiD.Cookie', function() {
 
     var assert = chai.assert;
-    var setup = {client_id : '4d00e8d6bf92fc8648000000', server: 'stage.payment.schibsted.se', useSessionCluster:false, logging:false};
-    var setupProd = {client_id : '4d00e8d6bf92fc8648000000', server: 'payment.schibsted.se', logging:false, cookie: false, cache:false};
+    var setup = {client_id : '4d00e8d6bf92fc8648000000', server: 'stage.payment.schibsted.se', useSessionCluster:false, logging:false, setVarnishCookie: true};
+    var setupProd = {client_id : '4d00e8d6bf92fc8648000000', server: 'payment.schibsted.se', logging:false, cookie: false, cache:false, setVarnishCookie: true};
     var spidCookie = require('../../src/spid-cookie'),
         SPiD = require('../../src/spid-sdk');
 
@@ -44,7 +44,7 @@ describe('SPiD.Cookie', function() {
 
         it('SPiD.Cookie.setVarnishCookie should set varnish cookie', function() {
             var session = {user:123, sp_id: 123, expiresIn: 5000, baseDomain: cookieDomain};
-            spidCookie.setVarnishCookie(session, session.expiresIn);
+            spidCookie.tryVarnishCookie(session, session.expiresIn);
 
             assert.notEqual(document.cookie.indexOf('SP_ID'), -1);
         });
